ODR/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.19
python-jose[cryptography]
